Fix watershed outline disappearing on watershed search and selection.
