+++ client/src/api/rhessysOutputsApi.ts
@@ -1,10 +1,10 @@
 import { queryOptions, keepPreviousData } from "@tanstack/react-query";
 import { API_ENDPOINTS } from "./apiEndpoints";
 import { checkResponse } from "./errors";
-import { postQuery, toFiniteNumber } from "./queryUtils";
+import { postQuery, toFiniteNumber } from "./utils";
 import { queryKeys } from "./queryKeys";
 
-import { getVariableMeta, resolveParquetConfig } from "./rhessys/utils";
+import { getVariableMeta, resolveParquetConfig } from "./utils";
 
 import type {
   RhessysOutputListResponse,
