+++ client/src/components/map/constants.ts
@@ -1,6 +1,7 @@
 export const defaultStyle = {
   color: "#4a83ec",
   weight: 3,
+  opacity: 1,
   fillColor: "#4a83ec",
   fillOpacity: 0.25,
 };
@@ -8,6 +9,7 @@ export const defaultStyle = {
 export const selectedStyle = {
   color: "#ffff00",
   weight: 3,
+  opacity: 1,
   fillOpacity: 0,
 };
